+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as UITable } from './UITable';
+export { default as UISearch } from './UISearch';
