+++ src/interfaces/Asset.ts
@@ -1,0 +1,5 @@
+export interface IAsset {
+    // name: string;
+    price: string;
+    exchange: string;
+}
