+++ src/interfaces/Table.ts
@@ -1,0 +1,20 @@
+import { IAsset } from './Asset'
+
+export type Order = 'asc' | 'desc';
+
+export interface HeadCell {
+    disablePadding: boolean;
+    id: keyof IAsset;
+    label: string;
+    numeric: boolean;
+    align: 'left' | 'right';
+}
+
+export interface EnhancedTableProps {
+    numSelected: number;
+    onRequestSort: (event: React.MouseEvent<unknown>, property: keyof IAsset) => void;
+    onSelectAllClick?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+    order: Order;
+    orderBy: string;
+    rowCount: number;
+}
